Revert "move workUnitAsyncStorage back into "client/components" (#71046)"
This reverts commit b6be4ad534ee95a944744d4865113d6da66c219d.
