+++ packages/next/src/server/app-render/work-unit-async-storage-instance.ts
@@ -1,4 +1,4 @@
-import { createAsyncLocalStorage } from './async-local-storage'
+import { createAsyncLocalStorage } from '../../client/components/async-local-storage'
 import type { WorkUnitAsyncStorage } from './work-unit-async-storage.external'
 
 export const workUnitAsyncStorage: WorkUnitAsyncStorage =
