+++ packages/next/src/server/node-environment-extensions/utils.tsx
@@ -1,5 +1,5 @@
 import { workAsyncStorage } from '../../client/components/work-async-storage.external'
-import { workUnitAsyncStorage } from '../../client/components/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import { abortOnSynchronousDynamicDataAccess } from '../app-render/dynamic-rendering'
 
 export function io(expression: string) {
