+++ packages/next/src/server/request/connection.ts
@@ -1,5 +1,5 @@
 import { workAsyncStorage } from '../../client/components/work-async-storage.external'
-import { workUnitAsyncStorage } from '../../client/components/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import {
   postponeWithTracking,
   throwToInterruptStaticGeneration,
