+++ packages/next/src/server/request/draft-mode.ts
@@ -1,9 +1,9 @@
-import { getExpectedRequestStore } from '../../client/components/work-unit-async-storage.external'
+import { getExpectedRequestStore } from '../../server/app-render/work-unit-async-storage.external'
 
 import type { DraftModeProvider } from '../../server/async-storage/draft-mode-provider'
 
 import { workAsyncStorage } from '../../client/components/work-async-storage.external'
-import { workUnitAsyncStorage } from '../../client/components/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import { trackDynamicDataAccessed } from '../app-render/dynamic-rendering'
 import { createDedupedByCallsiteServerErrorLoggerDev } from '../create-deduped-by-callsite-server-error-loger'
