+++ packages/next/src/server/use-cache/use-cache-wrapper.ts
@@ -14,8 +14,8 @@ import {
 
 import type { WorkStore } from '../../client/components/work-async-storage.external'
 import { workAsyncStorage } from '../../client/components/work-async-storage.external'
-import type { UseCacheStore } from '../../client/components/work-unit-async-storage.external'
-import { workUnitAsyncStorage } from '../../client/components/work-unit-async-storage.external'
+import type { UseCacheStore } from '../app-render/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import { runInCleanSnapshot } from '../app-render/clean-async-snapshot.external'
 
 import type { ClientReferenceManifest } from '../../build/webpack/plugins/flight-manifest-plugin'
