+++ packages/next/src/server/web/spec-extension/revalidate.ts
@@ -5,7 +5,7 @@ import {
   NEXT_CACHE_SOFT_TAG_MAX_LENGTH,
 } from '../../../lib/constants'
 import { workAsyncStorage } from '../../../client/components/work-async-storage.external'
-import { workUnitAsyncStorage } from '../../../client/components/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from '../../../server/app-render/work-unit-async-storage.external'
 
 /**
  * This function allows you to purge [cached data](https://nextjs.org/docs/app/building-your-application/caching) on-demand for a specific cache tag.
