+++ test/e2e/app-dir/app-external/app/async-storage/page.js
@@ -1,4 +1,4 @@
-import { workUnitAsyncStorage } from 'next/dist/client/components/work-unit-async-storage.external'
+import { workUnitAsyncStorage } from 'next/dist/server/app-render/work-unit-async-storage.external'
 
 export default async function Page() {
   // cookies is undefined if not set
